Spike aggressive cleanup of projects
This commit introduces an experimental flag to be used whenever all project configurations
are assumed to be declared properly as inputs of tasks or services. In this case, when
a project is no longer needed in the execution graph, we release some of its resources.

!!!THIS IS COMPLETELY EXPERIMENTAL, DRAGONS AHEAD, USE AT YOUR OWN RISK!!!

In settings, activate the preview feature:

enableFeaturePreview('AGGRESSIVE_PROJECT_UNLOADING')