Only initialize JANSI when native services are enabled
JANSI is being initialized, even though native
services have been disabled by the system property. This problem has been introduced by #17440.

As an additional problem, the nativeBaseDir has
not been set in this case, which caused jansi to
be initialized in the current working directory.

This PR fixes both problems:
- Do not initialize JANSI when native integrations are disabled.
- Always set the native base dir.